Crate cat_gateway

Catalyst Data Gateway

Modules

build_info 🔒

Binary build info

cardano 🔒

Logic for orchestrating followers

cli 🔒

CLI interpreter for the service

db 🔒

Database Interfaces

jinja 🔒

minijinja static global variables

logger 🔒

Setup for logging for the service.

metrics 🔒

This module contains submodules related to metrics report and analytics.

service 🔒

Main entrypoint to the service

settings 🔒

Command line and environment variable settings for the service

utils 🔒

General Purpose utility functions

Functions

main 🔒