cat_gateway
Module cardano

Logic for orchestrating followers

Modules

util 🔒

Block stream parsing and filtering utils

Structs

SyncParams 🔒

Data we return from a sync task.

SyncTask 🔒

The synchronisation task, and its state. There should ONLY ever be one of these at any time.

Constants

BACKOFF_RANGE_MULTIPLIER 🔒

The range we generate random backoffs within given a base backoff value.

INDEXING_DB_READY_WAIT_INTERVAL 🔒

How long we wait between checks for connection to the indexing DB to be ready.

MAX_BLOCKS_BATCH_LEN 🔒

Blocks batch length that will trigger the blocks buffer to be written to the database.

Functions

start_followers 🔒

Start followers as per defined in the config

start_sync_for 🔒

Start syncing a particular network

sync_subchain 🔒

Sync a portion of the blockchain. Set end to TIP_POINT to sync the tip continuously.