Module cat_gateway::cardano::cip36_registration

Verify registration TXs

Structs

Cip36Metadata 🔒

A catalyst CIP-36 registration on Cardano

Nonce 🔒

Nonce

PubKey 🔒

Pub key

Registration 🔒

CIP-36 registration info part

RewardsAddress 🔒

Rewards address

VotingPurpose 🔒

Voting purpose

Enums

VotingInfo 🔒

The source of voting power for a given registration

Functions

inspect_nonce 🔒

Extract Nonce

inspect_registration_from_metadata 🔒

Extract registration from tx metadata

inspect_rewards_addr 🔒

Extract and validate rewards address A Shelley payment address (see CIP-0019) discriminated for the same network this transaction is submitted to, to receive rewards.

inspect_signature 🔒

Extract signature

inspect_stake_key 🔒

Extract stake key.

inspect_voting_info 🔒

Extract voting info Voting key: simply an ED25519 public key. This is the spending credential in the side chain that will receive voting power from this delegation. For direct voting it’s necessary to have the corresponding private key to cast votes in the side chain

inspect_voting_purpose 🔒

Extract optional voting purpose

inspect_witness_from_metadata 🔒

Extract witness from tx metadata

is_valid_rewards_address 🔒

Reward addresses start with a single header byte identifying their type and the network, followed by 28 bytes of payload identifying either a stake key hash or a script hash. Function accepts this first header prefix byte. Validates first nibble is within the address range: 0x0? - 0x7? + 0xE? , 0xF? Validates second nibble matches network id: 0/1

original_61284_payload 🔒

Rebuild 61284 from pallas metadata to match original signed 61284 payload, pallas does not preserve exact 61284 bytes which is required for signature verification. It must be rebuilt and re-serialized to perform signature verification.

validate_cip36_registration 🔒

Validate binary data against CIP-36 registration CDDL spec

validate_cip36_witness 🔒

Validate binary data against CIP-36 registration CDDL spec

validate_signature

The signature is generated by:

Type Aliases

ErrorReport 🔒

Error report for serializing