cat_gateway::cardano

Module util

Block stream parsing and filtering utils

Structs

Asset

Assets

PolicyAsset

Parsed Assets

Functions

extract_stake_credentials_from_certs

Extract stake credentials from certificates. Stake credentials are 28 byte blake2b hashes.

find_matching_stake_credential

Match hashed witness pub keys with hashed stake credentials from the TX certificates to identify the correct stake credential key.

parse_child_assets 🔒

Parse child assets

parse_policy_assets 🔒

Extract assets

valid_era

Eras before staking should be ignored

Type Aliases

StakeCredentialHash

Stake credential hash from the certificate

StakeCredentialKey

Correct stake credential key in hex

WitnessHash

Witness pub key hashed with blake2b

WitnessPubKey

Witness pub key in hex