cat_gateway::db
Module event

Catalyst Election Database crate

Modules

error 🔒

Database Errors

legacy 🔒

Legacy event db queries and types

schema_check 🔒

Check if the schema is up-to-date.

Structs

EventDB 🔒

The Catalyst Event SQL Database

Enums

Error 🔒

EventDB Errors

Constants

DATABASE_SCHEMA_VERSION 🔒

Database version this crate matches. Must equal the last Migrations Version Number from event-db/migrations.

Statics

DEEP_QUERY_INSPECT 🔒

Is Deep Query Analysis enabled or not?

EVENT_DB_POOL 🔒

Postgres Connection Manager DB Pool Instance

Functions

establish_connection 🔒

Establish a connection to the database, and check the schema is up-to-date.

Type Aliases

SqlDbPool 🔒

Postgres Connection Manager DB Pool