cat_gateway::db::event::signed_docs
Function select_signed_docs

pub(crate) async fn select_signed_docs(
    id: &Uuid,
    ver: Option<&Uuid>,
) -> Result<FullSignedDoc>

Make a select query into the event-db by getting data from the signed_docs table.

This returns a single document. All data from the document is returned, including the payload and raw fields.

ver should be able to be optional, in which case get the latest ver of the given id.

Arguments:

id is a UUID v7

ver is a UUID v7