cat_gateway::db::index::block::certs

Struct CertInsertQuery

pub(crate) struct CertInsertQuery {
    stake_reg_data: Vec<StakeRegistrationInsertQuery>,
}

Insert Cert Queries

Fields
stake_reg_data: Vec<StakeRegistrationInsertQuery>

Stake Registration Data captured during indexing.

Implementations

impl CertInsertQuery

pub(crate) fn new() -> Self

Create new data set for Cert Insert Query Batch.

pub(crate) async fn prepare_batch( session: &Arc<Session>, cfg: &EnvVars, ) -> Result<SkipMap<u16, Arc<Batch>>>

Prepare Batch of Insert TXI Index Data Queries

fn stake_address( &mut self, cred: &StakeCredential, slot_no: Slot, txn: TxnIndex, register: bool, deregister: bool, delegation: Option<Vec<u8>>, block: &MultiEraBlock, )

Get the stake address for a hash, return an empty address if one can not be found.

fn index_alonzo_cert( &mut self, cert: &Certificate, slot: Slot, index: TxnIndex, block: &MultiEraBlock, )

Index an Alonzo Era certificate into the database.

fn index_conway_cert( &mut self, cert: &Certificate, slot_no: Slot, txn: TxnIndex, block: &MultiEraBlock, )

Index a certificate from a conway transaction.

pub(crate) fn index( &mut self, txs: &MultiEraTx<'_>, slot: Slot, index: TxnIndex, block: &MultiEraBlock, )

Index the certificates in a transaction.

pub(crate) fn execute( self, session: &Arc<CassandraSession>, ) -> Vec<JoinHandle<Result<Vec<QueryResult>>>>

Execute the Certificate Indexing Queries.

Consumes the self and returns a vector of futures.