cat_gateway::db::index

Module queries

Pre-prepare queries for a given session.

This improves query execution time.

Modules

purge 🔒

Queries for purging volatile data.

rbac 🔒

RBAC queries.

registrations 🔒

Registration related queries.

staked_ada 🔒

Staked ADA related queries.

sync_status 🔒

sync status update and query.

Structs

PreparedQueries 🔒

All prepared queries for a session.

Enums

PreparedQuery 🔒

All Prepared insert Queries that we know about.

PreparedSelectQuery 🔒

All prepared SELECT query statements (return data).

PreparedUpsertQuery 🔒

All prepared UPSERT query statements (inserts/updates a single value of data).

Functions

session_execute_batch 🔒

Execute a Batch query with the given parameters.

session_execute_iter 🔒

Executes a select query with the given parameters.

Type Aliases

FallibleQueryResults 🔒

A set of query responses that can fail.

FallibleQueryTasks 🔒

A set of query responses from tasks that can fail.

SizedBatch 🔒

Batches of different sizes, prepared and ready for use.