cat_gateway::db::index::queries
Module purge

Queries for purging volatile data.

Modules

chain_root_for_role0_key 🔒

Chain Root For Role0 Key (RBAC 509 registrations) Queries used in purging data.

chain_root_for_stake_address 🔒

Chain Root For Stake Address (RBAC 509 registrations) Queries used in purging data.

chain_root_for_txn_id 🔒

Chain Root For TX ID (RBAC 509 registrations) Queries used in purging data.

cip36_registration 🔒

CIP-36 registration Queries used in purging data.

cip36_registration_for_vote_key 🔒

CIP-36 registration by Vote Key Queries used in purging data.

cip36_registration_invalid 🔒

CIP-36 Registration (Invalid) Queries used in purging data.

rbac509_registration 🔒

RBAC 509 Registration Queries used in purging data.

stake_registration 🔒

Stake Registration Queries used in purging data.

txi_by_hash 🔒

Transaction Inputs (ADA or a native asset) queries used in purging data.

txo_ada 🔒

TXO by Stake Address Queries used in purging data.

txo_assets 🔒

TXO Assets by Stake Address Queries used in purging data.

unstaked_txo_ada 🔒

Unstaked Transaction Outputs (ADA), by their transaction hash, queries used in purging data.

unstaked_txo_assets 🔒

TXO Assets by TXN Hash Queries used in purging data.

Structs

PreparedQueries 🔒

All prepared purge queries for a session.

Enums

PreparedDeleteQuery 🔒

All prepared DELETE query statements (purge DB table rows).

PreparedSelectQuery 🔒

All prepared SELECT query statements (primary keys from table).

Constants

NO_PARAMS 🔒

No parameters