cat_gateway::db::index::queries::rbac

Module get_rbac_registrations

Get RBAC registrations by Catalyst ID.

Structs

Query 🔒

Get registrations by Catalyst ID query.

QueryParams 🔒

Get registrations by Catalyst ID query params.

Constants

QUERY 🔒

Get registrations by Catalyst ID query.

Functions

build_reg_chain 🔒

Build a registration chain from the given indexed data.

indexed_registrations 🔒

Returns a sorted list of all registrations for the given Catalyst ID from the database.

registration 🔒

A helper function to return a RBAC registration from the given block and slot.