Module cat_gateway::db::index::schema

Index Schema

Constants

CREATE_NAMESPACE_CQL 🔒

Keyspace Create (Templated)

CREATE_TABLE_TXI_BY_STAKE_ADDRESS_CQL 🔒

TXI by Stake Address Table Schema

CREATE_TABLE_TXO_ASSETS_BY_STAKE_ADDRESS_CQL 🔒

TXO Assets by Stake Address Table Schema

CREATE_TABLE_TXO_BY_STAKE_ADDRESS_CQL 🔒

TXO by Stake Address Table Schema

SCHEMA_VERSION 🔒

The version of the Schema we are using. Must be incremented if there is a breaking change in any schema tables below.

Functions

create_namespace 🔒

Create the namespace we will use for this session Ok to run this if the namespace already exists.

create_schema 🔒

Create the Schema on the connected Cassandra DB

create_txi_tables 🔒

Create tables for holding volatile TXI data

create_txo_tables 🔒

Create tables for holding TXO data.

namespace 🔒

Get the namespace for a particular db configuration