cat_gateway::db::index

Module session

Session creation and storage

Structs

CassandraSession 🔒

All interaction with cassandra goes through this struct.

Enums

CassandraSessionError 🔒

Represents errors that can occur while interacting with a Cassandra session.

CompressionChoice 🔒

Configuration Choices for compression

TlsChoice 🔒

Configuration Choices for TLS.

Statics

INIT_SESSION_ERROR 🔒

Session error while initialization.

PERSISTENT_SESSION 🔒

Persistent DB Session.

VOLATILE_SESSION 🔒

Volatile DB Session.

Functions

make_execution_profile 🔒

Create a new execution profile based on the given configuration.

make_session 🔒

Construct a session based on the given configuration.

retry_init 🔒

Continuously try and init the DB, if it fails, backoff.