cat_gateway::db

Module types

Wrappers over commonly used types that can be stored to and load from a database.

Re-exports

pub use catalyst_id::DbCatalystId;

pub use public_key::DbPublicKey;

pub use slot::DbSlot;

pub use stake_address::DbStakeAddress;

pub use transaction_id::DbTransactionId;

pub use transaction_index::DbTxnIndex;

pub use txn_output_offset::DbTxnOutputOffset;

pub use uuidv4::DbUuidV4;

Modules

catalyst_id 🔒

A IdUri wrapper that can be stored to and load from a database.

public_key 🔒

A VerifyingKey wrapper that can be stored to and load from a database.

slot 🔒

A Slot wrapper that can be stored to and load from a database.

stake_address 🔒

A binary CIP-19 stack address (29 bytes) that can be stored to and load from a database.

transaction_id 🔒

A TransactionHash hash wrapper that can be stored to and load from a database.

transaction_index 🔒

A TxnIndex wrapper that can be stored to and load from a database.

txn_output_offset 🔒

A TxnOutputOffset wrapper that can be stored to and load from a database.

uuidv4 🔒

A UuidV4 wrapper that can be stored to and load from a database.