Struct cat_gateway::event_db::EventDB

pub(crate) struct EventDB {
    pool: Pool<PostgresConnectionManager<NoTls>>,
}

Connection to the Election Database

Fields
pool: Pool<PostgresConnectionManager<NoTls>>

Internal database connection. DO NOT MAKE PUBLIC. All database operations (queries, inserts, etc) should be constrained to this crate and should be exported with a clean data access api.

Implementations

impl EventDB

pub(crate) async fn get_follower_config( &self ) -> Result<Vec<FollowerConfig>, Error>

Config query

impl EventDB

pub(crate) async fn index_follower_data( &self, slot_no: i64, network: Network, epoch_no: i64, block_time: DateTime<Utc>, block_hash: String ) -> Result<(), Error>

Index follower block stream

pub(crate) async fn get_slot_info( &self, date_time: DateTime<Utc>, network: Network, query_type: SlotInfoQueryType ) -> Result<(i64, String, DateTime<Utc>), Error>

Get slot info for the provided date-time and network and query type

pub(crate) async fn last_updated_metadata( &self, network: Network ) -> Result<(i64, String, DateTime<Utc>), Error>

Check when last update occurred. Start follower from where previous follower left off.

pub(crate) async fn refresh_last_updated( &self, last_updated: DateTime<Utc>, slot_no: i64, block_hash: String, network: Network, machine_id: &String ) -> Result<(), Error>

Mark point in time where the last follower finished indexing in order for future followers to pick up from this point

impl EventDB

const BALLOTS_VOTE_OPTIONS_PER_EVENT_QUERY: &'static str = "SELECT vote_options.objective, proposal.id as proposal_id, objective.id as objective_id FROM proposal INNER JOIN objective ON proposal.objective = objective.row_id INNER JOIN vote_options ON objective.vote_options = vote_options.id WHERE objective.event = $1;"

Ballot vote options per event query template

const BALLOTS_VOTE_OPTIONS_PER_OBJECTIVE_QUERY: &'static str = "SELECT vote_options.objective, proposal.id as proposal_id FROM proposal INNER JOIN objective ON proposal.objective = objective.row_id INNER JOIN vote_options ON objective.vote_options = vote_options.id WHERE objective.event = $1 AND objective.id = $2;"

Ballot vote options per objective query template

const BALLOT_VOTE_OPTIONS_QUERY: &'static str = "SELECT vote_options.objective FROM proposal INNER JOIN objective ON proposal.objective = objective.row_id INNER JOIN vote_options ON objective.vote_options = vote_options.id WHERE objective.event = $1 AND objective.id = $2 AND proposal.id = $3;"

Ballot vote options query template

const BALLOT_VOTE_PLANS_QUERY: &'static str = "SELECT proposal_voteplan.bb_proposal_index, voteplan.id, voteplan.category, voteplan.encryption_key, voteplan.group_id FROM proposal_voteplan INNER JOIN proposal ON proposal_voteplan.proposal_id = proposal.row_id INNER JOIN voteplan ON proposal_voteplan.voteplan_id = voteplan.row_id INNER JOIN objective ON proposal.objective = objective.row_id WHERE objective.event = $1 AND objective.id = $2 AND proposal.id = $3;"

Ballot vote plans query template

impl EventDB

pub(crate) async fn get_ballot( &self, event: EventId, objective: ObjectiveId, proposal: ProposalId ) -> Result<Ballot, Error>

Get ballot query

pub(crate) async fn get_objective_ballots( &self, event: EventId, objective: ObjectiveId ) -> Result<Vec<ProposalBallot>, Error>

Get objective’s ballots query

pub(crate) async fn get_event_ballots( &self, event: EventId ) -> Result<Vec<ObjectiveBallots>, Error>

Get event’s ballots query

impl EventDB

const OBJECTIVES_QUERY: &'static str = "SELECT objective.row_id, objective.id, objective.title, objective.description, objective.deleted, objective.rewards_currency, objective.rewards_total, objective.extra, objective_category.name, objective_category.description as objective_category_description FROM objective INNER JOIN objective_category on objective.category = objective_category.name WHERE objective.event = $1 LIMIT $2 OFFSET $3;"

Objectives query template

const VOTING_GROUPS_QUERY: &'static str = "SELECT voteplan.group_id as group, voteplan.token_id as voting_token FROM voteplan WHERE objective_id = $1;"

Voting Groups query template

impl EventDB

pub(crate) async fn get_objectives( &self, event: EventId, limit: Option<i64>, offset: Option<i64> ) -> Result<Vec<Objective>, Error>

Get objectives query

impl EventDB

const PROPOSALS_QUERY: &'static str = "SELECT proposal.id, proposal.title, proposal.summary, proposal.deleted FROM proposal INNER JOIN objective on proposal.objective = objective.row_id WHERE objective.event = $1 AND objective.id = $2 LIMIT $3 OFFSET $4;"

Proposals query template

const PROPOSAL_QUERY: &'static str = "SELECT proposal.id, proposal.title, proposal.summary, proposal.deleted, proposal.extra, proposal.funds, proposal.url, proposal.files_url, proposal.proposer_name, proposal.proposer_contact, proposal.proposer_url, proposal.public_key FROM proposal INNER JOIN objective on proposal.objective = objective.row_id WHERE objective.event = $1 AND objective.id = $2 AND proposal.id = $3;"

Proposal details query template

impl EventDB

pub(crate) async fn get_proposal( &self, event: EventId, objective: ObjectiveId, proposal: ProposalId ) -> Result<Proposal, Error>

Get proposal query

pub(crate) async fn get_proposals( &self, event: EventId, objective: ObjectiveId, limit: Option<i64>, offset: Option<i64> ) -> Result<Vec<ProposalSummary>, Error>

Get proposals query

impl EventDB

const RATINGS_PER_REVIEW_QUERY: &'static str = "SELECT review_rating.metric, review_rating.rating, review_rating.note FROM review_rating WHERE review_rating.review_id = $1;"

Rating per review query template

const REVIEWS_QUERY: &'static str = "SELECT proposal_review.row_id, proposal_review.assessor FROM proposal_review INNER JOIN proposal on proposal.row_id = proposal_review.proposal_id INNER JOIN objective on proposal.objective = objective.row_id WHERE objective.event = $1 AND objective.id = $2 AND proposal.id = $3 LIMIT $4 OFFSET $5;"

Review query template

const REVIEW_TYPES_QUERY: &'static str = "SELECT review_metric.row_id, review_metric.name, review_metric.description, review_metric.min, review_metric.max, review_metric.map, objective_review_metric.note, objective_review_metric.review_group FROM review_metric INNER JOIN objective_review_metric on review_metric.row_id = objective_review_metric.metric INNER JOIN objective on objective_review_metric.objective = objective.row_id WHERE objective.event = $1 AND objective.id = $2 LIMIT $3 OFFSET $4;"

Review types query template

impl EventDB

pub(crate) async fn get_reviews( &self, event: EventId, objective: ObjectiveId, proposal: ProposalId, limit: Option<i64>, offset: Option<i64> ) -> Result<Vec<AdvisorReview>, Error>

Get reviews query

pub(crate) async fn get_review_types( &self, event: EventId, objective: ObjectiveId, limit: Option<i64>, offset: Option<i64> ) -> Result<Vec<ReviewType>, Error>

Get review types query

impl EventDB

const EVENTS_QUERY: &'static str = "SELECT event.row_id, event.name, event.start_time, event.end_time, snapshot.last_updated FROM event LEFT JOIN snapshot ON event.row_id = snapshot.event ORDER BY event.row_id ASC LIMIT $1 OFFSET $2;"

Events query template

const EVENT_GOALS_QUERY: &'static str = "SELECT goal.idx, goal.name FROM goal WHERE goal.event_id = $1;"

Event goals query template

const EVENT_QUERY: &'static str = "SELECT event.row_id, event.name, event.start_time, event.end_time, event.snapshot_start, event.registration_snapshot_time, event.voting_power_threshold, event.max_voting_power_pct, event.insight_sharing_start, event.proposal_submission_start, event.refine_proposals_start, event.finalize_proposals_start, event.proposal_assessment_start, event.assessment_qa_start, event.voting_start, event.voting_end, event.tallying_end, snapshot.last_updated FROM event LEFT JOIN snapshot ON event.row_id = snapshot.event WHERE event.row_id = $1;"

Event details query template

impl EventDB

pub(crate) async fn get_events( &self, limit: Option<i64>, offset: Option<i64> ) -> Result<Vec<EventSummary>, Error>

Get events query

pub(crate) async fn get_event(&self, event: EventId) -> Result<Event, Error>

Get event query

impl EventDB

const DELEGATIONS_BY_EVENT_QUERY: &'static str = "SELECT contribution.voting_key, contribution.voting_group, contribution.voting_weight, contribution.value, contribution.reward_address FROM contribution INNER JOIN snapshot ON contribution.snapshot_id = snapshot.row_id WHERE contribution.stake_public_key = $1 AND snapshot.event = $2;"

Delegations by event query template

const DELEGATOR_SNAPSHOT_INFO_BY_EVENT_QUERY: &'static str = " SELECT snapshot.as_at, snapshot.last_updated, snapshot.final FROM snapshot WHERE snapshot.event = $1 LIMIT 1;"

Delegator snapshot info by event query template

const DELEGATOR_SNAPSHOT_INFO_BY_LAST_EVENT_QUERY: &'static str = "SELECT snapshot.event, snapshot.as_at, snapshot.last_updated, snapshot.final FROM snapshot WHERE snapshot.last_updated = (SELECT MAX(snapshot.last_updated) as last_updated from snapshot) LIMIT 1;"

Delegator snapshot info by last event query template

const TOTAL_BY_EVENT_VOTING_QUERY: &'static str = "SELECT SUM(voter.voting_power)::BIGINT as total_voting_power FROM voter INNER JOIN snapshot ON voter.snapshot_id = snapshot.row_id WHERE voter.voting_group = $1 AND snapshot.event = $2;"

Total By Event Query template

const TOTAL_BY_LAST_EVENT_VOTING_QUERY: &'static str = "SELECT SUM(voter.voting_power)::BIGINT as total_voting_power FROM voter INNER JOIN snapshot ON voter.snapshot_id = snapshot.row_id AND snapshot.last_updated = (SELECT MAX(snapshot.last_updated) as last_updated from snapshot) WHERE voter.voting_group = $1;"

Total By Last Event Query template

const TOTAL_POWER_BY_EVENT_QUERY: &'static str = "SELECT SUM(voter.voting_power)::BIGINT as total_voting_power FROM voter INNER JOIN snapshot ON voter.snapshot_id = snapshot.row_id WHERE snapshot.event = $1;"

Total voting power by event query template

const TOTAL_POWER_BY_LAST_EVENT_QUERY: &'static str = "SELECT SUM(voter.voting_power)::BIGINT as total_voting_power FROM voter INNER JOIN snapshot ON voter.snapshot_id = snapshot.row_id WHERE snapshot.last_updated = (SELECT MAX(snapshot.last_updated) as last_updated from snapshot);"

Total voting power by last event query template

const VOTER_BY_EVENT_QUERY: &'static str = "SELECT voter.voting_key, voter.voting_group, voter.voting_power, snapshot.as_at, snapshot.last_updated, snapshot.final, SUM(contribution.value)::BIGINT as delegations_power, COUNT(contribution.value) AS delegations_count FROM voter INNER JOIN snapshot ON voter.snapshot_id = snapshot.row_id INNER JOIN contribution ON contribution.snapshot_id = snapshot.row_id WHERE voter.voting_key = $1 AND contribution.voting_key = $1 AND snapshot.event = $2 GROUP BY voter.voting_key, voter.voting_group, voter.voting_power, snapshot.as_at, snapshot.last_updated, snapshot.final;"

Voter By Event Query template

const VOTER_BY_LAST_EVENT_QUERY: &'static str = "SELECT snapshot.event, voter.voting_key, voter.voting_group, voter.voting_power, snapshot.as_at, snapshot.last_updated, snapshot.final, SUM(contribution.value)::BIGINT as delegations_power, COUNT(contribution.value) AS delegations_count FROM voter INNER JOIN snapshot ON voter.snapshot_id = snapshot.row_id INNER JOIN contribution ON contribution.snapshot_id = snapshot.row_id WHERE voter.voting_key = $1 AND contribution.voting_key = $1 AND snapshot.last_updated = (SELECT MAX(snapshot.last_updated) as last_updated from snapshot) GROUP BY snapshot.event, voter.voting_key, voter.voting_group, voter.voting_power, snapshot.as_at, snapshot.last_updated, snapshot.final;"

Voter By Last Event Query template

const VOTER_DELEGATORS_LIST_QUERY: &'static str = "SELECT contribution.stake_public_key FROM contribution INNER JOIN snapshot ON contribution.snapshot_id = snapshot.row_id WHERE contribution.voting_key = $1 AND snapshot.event = $2;"

Voter Delegators List Query template

impl EventDB

pub(crate) async fn get_voter( &self, event: &Option<EventId>, voting_key: String, with_delegations: bool ) -> Result<Voter, Error>

Get voter query

pub(crate) async fn get_delegator( &self, event: &Option<EventId>, stake_public_key: String ) -> Result<Delegator, Error>

Get delegator query

impl EventDB

const SEARCH_EVENTS_QUERY: &'static str = "SELECT event.row_id, event.name, event.start_time, event.end_time, snapshot.last_updated FROM event LEFT JOIN snapshot ON event.row_id = snapshot.event"

Search for events query template

const SEARCH_OBJECTIVES_QUERY: &'static str = "SELECT objective.id, objective.title, objective.description, objective.deleted, objective_category.name, objective_category.description as objective_category_description FROM objective INNER JOIN objective_category on objective.category = objective_category.name"

Search for objectives query template

const SEARCH_PROPOSALS_QUERY: &'static str = "SELECT DISTINCT proposal.id, proposal.title, proposal.summary, proposal.deleted FROM proposal"

Search for proposals query template

fn build_where_clause(table: &str, filter: &[SearchConstraint]) -> String

Build a where clause

fn build_order_by_clause(table: &str, order_by: &[SearchOrderBy]) -> String

Build an order by clause

fn construct_query(search_query: &SearchQuery) -> String

Construct a search query

fn construct_count_query(search_query: &SearchQuery) -> String

Construct a count query

async fn search_total( &self, search_query: SearchQuery, limit: Option<i64>, offset: Option<i64> ) -> Result<SearchResult, Error>

Search for a total.

async fn search_events( &self, search_query: SearchQuery, limit: Option<i64>, offset: Option<i64> ) -> Result<SearchResult, Error>

Search for events

async fn search_objectives( &self, search_query: SearchQuery, limit: Option<i64>, offset: Option<i64> ) -> Result<SearchResult, Error>

Search for objectives

async fn search_proposals( &self, search_query: SearchQuery, limit: Option<i64>, offset: Option<i64> ) -> Result<SearchResult, Error>

Search for proposals

impl EventDB

pub(crate) async fn search( &self, search_query: SearchQuery, total: bool, limit: Option<i64>, offset: Option<i64> ) -> Result<SearchResult, Error>

Search query

impl EventDB

const FUND_CHALLENGES_QUERY: &'static str = "SELECT objective.row_id AS internal_id, objective.id AS id, objective.category AS challenge_type, objective.title AS title, objective.description AS description, objective.rewards_total AS rewards_total, objective.proposers_rewards AS proposers_rewards, (objective.extra->'url') #>> '{}' AS challenge_url, (objective.extra->'sponsor') #>> '{}' AS highlights FROM objective WHERE objective.event = $1;"

const FUND_GOALS_QUERY: &'static str = "SELECT id, name AS goal_name FROM goal WHERE goal.event_id = $1;"

const FUND_GROUPS_QUERY: &'static str = "SELECT voteplan.token_id AS token_identifier, voting_group.name AS group_id FROM voting_group INNER JOIN voteplan ON voteplan.group_id = voting_group.name INNER JOIN objective ON voteplan.objective_id = objective.row_id WHERE objective.event = $1;"

const FUND_QUERY: &'static str = "SELECT this_fund.row_id AS id, this_fund.name AS fund_name, this_fund.description AS fund_goal, this_fund.registration_snapshot_time AS registration_snapshot_time, this_fund.voting_power_threshold AS voting_power_threshold, this_fund.start_time AS fund_start_time, this_fund.end_time AS fund_end_time, this_fund.insight_sharing_start AS insight_sharing_start, this_fund.proposal_submission_start AS proposal_submission_start, this_fund.refine_proposals_start AS refine_proposals_start, this_fund.finalize_proposals_start AS finalize_proposals_start, this_fund.proposal_assessment_start AS proposal_assessment_start, this_fund.assessment_qa_start AS assessment_qa_start, this_fund.snapshot_start AS snapshot_start, this_fund.voting_start AS voting_start, this_fund.voting_end AS voting_end, this_fund.tallying_end AS tallying_end, (this_fund.extra->'url'->'results') #>> '{}' AS results_url, (this_fund.extra->'url'->'survey') #>> '{}' AS survey_url, next_fund.row_id AS next_id, next_fund.start_time AS next_fund_start_time, next_fund.registration_snapshot_time AS next_registration_snapshot_time, next_fund.name AS next_fund_name, next_fund.insight_sharing_start AS next_insight_sharing_start, next_fund.proposal_submission_start AS next_proposal_submission_start, next_fund.refine_proposals_start AS next_refine_proposals_start, next_fund.finalize_proposals_start AS next_finalize_proposals_start, next_fund.proposal_assessment_start AS next_proposal_assessment_start, next_fund.assessment_qa_start AS next_assessment_qa_start, next_fund.snapshot_start AS next_snapshot_start, next_fund.voting_start AS next_voting_start, next_fund.voting_end AS next_voting_end, next_fund.tallying_end AS next_tallying_end FROM event this_fund LEFT JOIN event next_fund ON next_fund.row_id = this_fund.row_id + 1 WHERE this_fund.end_time > CURRENT_TIMESTAMP AT TIME ZONE 'UTC' AND this_fund.start_time < CURRENT_TIMESTAMP AT TIME ZONE 'UTC' ORDER BY this_fund.row_id DESC LIMIT 1;"

const FUND_VOTE_PLANS_QUERY: &'static str = "SELECT voteplan.row_id AS id, voteplan.id AS chain_voteplan_id, voteplan.category AS chain_voteplan_payload, voteplan.encryption_key AS chain_vote_encryption_key, event.row_id AS fund_id, voteplan.token_id AS token_identifier FROM voteplan INNER JOIN objective ON voteplan.objective_id = objective.row_id INNER JOIN event ON objective.event = event.row_id WHERE event.row_id = $1;"

impl EventDB

pub(crate) async fn get_fund(&self) -> Result<FundWithNext, Error>

Get fund query

impl EventDB

pub(crate) async fn schema_version_check(&self) -> Result<i32, Error>

Check the schema version. return the current schema version if its current. Otherwise return an error.

impl EventDB

pub(crate) async fn index_utxo_data( &self, txs: Vec<MultiEraTx<'_>>, slot_no: i64, network: Network ) -> Result<(), Error>

Index utxo data

pub(crate) async fn index_txn_data( &self, tx_id: &[u8], slot_no: i64, network: Network ) -> Result<(), Error>

Index txn metadata

pub(crate) async fn total_utxo_amount( &self, stake_credential: &'_ [u8], network: Network, slot_num: i64 ) -> Result<(i64, i64), Error>

Get total utxo amount

impl EventDB

async fn insert_voter_registration( &self, tx_id: String, stake_credential: &'_ [u8], public_voting_key: &'_ [u8], payment_address: &'_ [u8], metadata_cip36: &'_ [u8], nonce: i64, report: Value, valid: bool ) -> Result<(), Error>

Inserts voter registration data, replacing any existing data.

pub async fn index_registration_data( &self, txs: Vec<MultiEraTx<'_>>, slot_no: i64, network: Network ) -> Result<(), Error>

Index registration data