cat_gateway::metrics::chain_follower
Module reporter

All the related Chain Follower reporting metrics to the Prometheus service are inside this module.

Constants

FOLLOWER_METRIC_LABELS 🔒

Labels for the chain follower metrics

FOLLOWER_METRIC_NETWORKS 🔒

Chain networks use as the metrics namespace.

Statics

LIVE_BACKFILL_END 🔒

When backfill ended

LIVE_BACKFILL_FAILURES 🔒

Backfill Failures

LIVE_BACKFILL_FAILURE_TIME 🔒

The time of the last backfill failure

LIVE_BACKFILL_SIZE 🔒

Backfill size to achieve synchronization. (0 before sync completed)

LIVE_BACKFILL_START 🔒

When backfill started

LIVE_BLOCKS 🔒

Current Number of Live Blocks

LIVE_CONNECTED 🔒

Is there an active connection to the node

LIVE_FOLLOWER_COUNT 🔒

Number of active Followers

LIVE_HEAD_SLOT 🔒

The current head of the live chain slot#

LIVE_INVALID_BLOCKS 🔒

Blocks that failed to deserialize from the blockchain.

LIVE_LAST_CONNECT 🔒

Last reconnect time

LIVE_LAST_DISCONNECT 🔒

Last disconnect time

LIVE_NEW_BLOCKS 🔒

New blocks read from blockchain.

LIVE_RECONNECTS 🔒

Number of times we connected/re-connected to the Node.

LIVE_ROLLBACKS_FOLLOWER_COUNT 🔒

Rollbacks synthesized for followers.

LIVE_ROLLBACKS_LIVE_COUNT 🔒

Rollbacks we did on our live-chain in memory.

LIVE_ROLLBACKS_PEER_COUNT 🔒

Rollbacks reported by the Peer Node.

LIVE_SYNC_END 🔒

The Time that synchronization to this blockchain was complete up-to-tip. None = Not yet synchronized.

LIVE_SYNC_START 🔒

The Time that synchronization to this blockchain started

LIVE_TIP 🔒

The current live tip slot# as reported by the peer.

MITHRIL_CHANGED 🔒

Number of changed files from previous snapshot.

MITHRIL_DEDUPLICATED 🔒

Number of identical files deduplicated from previous snapshot.

MITHRIL_DEDUPLICATED_SIZE 🔒

Deduplicated Size vs previous snapshot.

MITHRIL_DL_END 🔒

Time we finished downloading the current snapshot.

MITHRIL_DL_FAILURES 🔒

Number of times download failed.

MITHRIL_DL_SIZE 🔒

Size of the download archive.

MITHRIL_DL_START 🔒

Time we started downloading the current snapshot.

MITHRIL_DOWNLOAD_OR_VALIDATION_FAILED 🔒

Download or Validation Failed.

MITHRIL_EXTRACT_END 🔒

Extraction end time.

MITHRIL_EXTRACT_FAILURES 🔒

Number of times extraction failed.

MITHRIL_EXTRACT_SIZE 🔒

Size of last extracted snapshot.

MITHRIL_EXTRACT_START 🔒

Extraction start time.

MITHRIL_FAILED_TO_ACTIVATE_NEW_SNAPSHOT 🔒

Failed to activate new snapshot.

MITHRIL_FAILED_TO_GET_TIP 🔒

Failed to get tip from Mithril snapshot.

MITHRIL_INVALID_BLOCKS 🔒

Blocks that failed to deserialize from the Mithril immutable chain.

MITHRIL_LAST_DL_DURATION 🔒

Time the last download took.

MITHRIL_NEW 🔒

Number of new files from previous snapshot.

MITHRIL_TIP 🔒

Immutable TIP Slot# - Origin = No downloaded snapshot.

MITHRIL_TIP_DID_NOT_ADVANCE 🔒

Tip failed to advance.

MITHRIL_TIP_FAILED_TO_SEND_TO_UPDATER 🔒

Failed to send new tip to updater.

MITHRIL_UPDATES 🔒

Number of Mithril Snapshots that have downloaded successfully.

MITHRIL_VALIDATE_END 🔒

Mithril Certificate Validation End Time.

MITHRIL_VALIDATE_FAILURES 🔒

Number of times validation failed (bad snapshot).

MITHRIL_VALIDATE_START 🔒

Mithril Certificate Validation Start Time.