cat_gateway::metrics::chain_indexer

Module reporter

All the related Chain Indexer reporting metrics to the Prometheus service are inside this module.

Constants

METRIC_LABELS 🔒

Labels for the metrics.

Statics

CURRENT_IMMUTABLE_TIP_SLOT 🔒

Chain Indexer current immutable tip slot#.

CURRENT_LIVE_TIP_SLOT 🔒

Chain Indexer current live tip slot#.

HIGHEST_COMPLETE_INDEXED_SLOT 🔒

Chain Indexer highest complete indexed slot#.

PURGED_SLOTS 🔒

Chain Indexer number of purged slots.

REACHED_TIP 🔒

Chain Indexer indicator whether tip is reached or in progress.

RUNNING_INDEXER_TASKS_COUNT 🔒

Chain Indexer number of current running indexer tasks.

TRIGGERED_BACKWARD_PURGES_COUNT 🔒

Chain Indexer number of times triggering backward data purge.

TRIGGERED_FORWARD_PURGES_COUNT 🔒

Chain Indexer number of times triggering forward data purge.