cat_gateway::metrics

Module memory

Metrics related to memory analytics.

Modules

reporter 🔒

All the related memory reporting metrics to the Prometheus service are inside this module.

Statics

GLOBAL 🔒

Use the instrumented allocator for gathering allocation statistics. Note: This wraps the global allocator. All structs that use the global allocator can be tracked.

IS_INITIALIZED 🔒

This is to prevent the init function from accidentally being called multiple times.

Functions

init_metrics_reporter 🔒

Starts a background thread to periodically update memory metrics.