cat_gateway::metrics::memory

Module reporter

All the related memory reporting metrics to the Prometheus service are inside this module.

Constants

MEMORY_METRIC_LABELS 🔒

Labels for the memory metrics

Statics

MEMORY_ALLOCATION_COUNT 🔒

The number of allocation count in the heap.

MEMORY_BYTES_ALLOCATED 🔒

The amount of accumulative allocated bytes in the heap.

MEMORY_BYTES_DEALLOCATED 🔒

The amount of accumulative deallocated bytes in the heap.

MEMORY_BYTES_REALLOCATED 🔒

The amount of accumulative reallocated bytes in the heap.

MEMORY_DEALLOCATION_COUNT 🔒

The number of deallocation count in the heap.

MEMORY_PHYSICAL_USAGE 🔒

The “physical” memory used by this process, in bytes.

MEMORY_REALLOCATION_COUNT 🔒

The number of reallocation count in the heap.

MEMORY_VIRTUAL_USAGE 🔒

The “virtual” memory used by this process, in bytes.