Module cat_gateway::registration

Verify registration TXs

Structs

CddlConfig

Cddl schema: https://cips.cardano.org/cips/cip36/schema.cddl

Nonce

Nonce

PubKey

Pub key

Registration

A catalyst registration on Cardano in either CIP-15 or CIP-36 format

RewardsAddress

Rewards address

Signature

Ed25519 signature.

VotingPurpose

Voting purpose

Enums

VotingInfo

The source of voting power for a given registration

Constants

CIP36_61284 🔒

https://cips.cardano.org/cips/cip36/schema.cddl

CIP36_61285 🔒

https://cips.cardano.org/cips/cip36/schema.cddl

COMPONENT_SIZE 🔒

Size of a single component of an Ed25519 signature.

DELEGATIONS_OR_DIRECT 🔒

Cip36

KEY_61284 🔒

Cip36 - 61284 entries

NETWORK_ID 🔒

Networks

NONCE 🔒

Cip36

PAYMENT_ADDRESS 🔒

Cip36

STAKE_ADDRESS 🔒

Cip36

VOTE_KEY 🔒

Cip36

VOTE_PURPOSE 🔒

Cip36

WEIGHT 🔒

Cip36

Functions

inspect_metamap_reg

Parse cip36 registration tx

inspect_nonce

Extract Nonce

inspect_rewards_addr

Extract and validate rewards address

inspect_stake_key

Extract stake key

inspect_voting_key

Extract voting key

inspect_voting_purpose

Extract optional voting purpose

is_valid_rewards_address

Reward addresses start with a single header byte identifying their type and the network, followed by 28 bytes of payload identifying either a stake key hash or a script hash. Function accepts this first header prefix byte. Validates first nibble is within the address range: 0x0? - 0x7? + 0xE? , 0xF? Validates second nibble matches network id: 0/1

parse_registrations_from_metadata

Extract registrations information for TX metadata Collect secondary errors for granular json error report

raw_sig_conversion

Convert raw 61285 cbor to witness signature

validate_reg_cddl

Validate raw registration binary against 61284 CDDL spec

Type Aliases

ComponentBytes

Size of an R or s component of an Ed25519 signature when serialized as bytes.

ErrorReport

Error report for serializing

SignatureBytes

Ed25519 signature serialized as a byte array.