Module cat_gateway::service::api::auth::token

Token encoding decoding logic

Structs

Kid

Key ID - Blake2b-128 hash of the Role 0 Certificate defining the Session public key. BLAKE2b-128 produces digest side of 16 bytes.

SignatureEd25519

Ed25519 signatures are (64 bytes)

UlidBytes

Identifier for this token, encodes both the time the token was issued and a random nonce.

Functions

decode_auth_token_ed25519

Decode base64 cbor encoded auth token into constituent parts of (kid, ulid, signature) e.g catv1.UAARIjNEVWZ3iJmqu8zd7v9QAZEs7HHPLEwUpV1VhdlNe1hAAAAAAAAAAAAA…

encode_auth_token_ed25519

The Encoded Binary Auth Token is a [CBOR sequence] that consists of 3 fields [ kid, ulid, signature ]. ED25519 Signature over the preceding two fields - sig(cbor(kid), cbor(ulid))