cat_gateway::service::api::cardano::cip36
Module filter

Implementation of the GET /cardano/cip36 endpoint

Functions

check_stake_addr_voting_key_association 🔒

Stake addresses need to be individually checked to make sure they are still actively associated with the voting key, and have not been registered to another voting key.

cross_reference_key 🔒

Check associated voting key matches registration voting key

get_all_invalid_registrations 🔒

Get all invalid registrations

get_all_registrations

Get all cip36 registrations.

get_all_registrations_from_stake_pub_key 🔒

Get all cip36 registrations for a given stake address.

get_invalid_registrations 🔒

Get invalid registrations for stake addr after given slot#

get_registration_from_stake_addr

Get registration from stake addr

get_registration_given_slot_no 🔒

Get registration given slot#

get_registration_given_stake_key_hash 🔒

Get registration given a stake key hash, it can be time specific based on asat param, or the latest registration returned if no asat given.

get_registration_given_vote_key 🔒

Get registration given a vote key, time specific based on asat param, or latest registration returned if no asat given.

invalid_filter 🔒

Filter out any invalid registrations that occurred before this Slot no

latest_vote_key 🔒

Stake addr may have multiple registrations and multiple vote key associations, filter out latest vote key.

slot_filter 🔒

Filter out any registrations that occurred after this Slot no

snapshot

ALL Get all registrations or constrain if slot# given.

sort_latest_registration 🔒

Sort latest registrations for a given stake address, sort by slot no and return latest.