cat_gateway::service::api::cardano::cip36::filter

Function get_registration_given_stake_key_hash

pub(crate) async fn get_registration_given_stake_key_hash(
    stake_address: StakeAddress,
    session: Arc<CassandraSession>,
    asat: Option<SlotNo>,
) -> WithErrorResponses<Cip36Registration>

Get registration given a stake key hash, it can be time specific based on asat param, or the latest registration returned if no asat given.