cat_gateway::service::api::cardano::cip36
Module old_endpoint

Implementation of the GET /cardano/cip36 endpoint

Enums

ResponseMultipleRegistrations 🔒

Endpoint responses.

ResponseSingleRegistration 🔒

Endpoint responses.

Functions

check_stake_addr_voting_key_association 🔒

Stake addresses need to be individually checked to make sure they are still actively associated with the voting key, and have not been registered to another voting key.

get_all_registrations_from_stake_pub_key 🔒

Get all cip36 registrations for a given stake address.

get_associated_vote_key_registrations 🔒

Returns the list of stake address registrations currently associated with a given voting key and returns any erroneous registrations which occur AFTER the slot# of the last valid registration.

get_invalid_registrations 🔒

Get invalid registrations for stake addr after given slot no

get_latest_registration_from_stake_addr 🔒

Get latest registration given a stake public key

get_latest_registration_from_stake_key_hash 🔒

Get latest registration given a stake key hash.

latest_registration_from_stake_addr 🔒

Get latest registration given a stake addr

sort_latest_registration 🔒

Sort latest registrations for a given stake address sorting by slot no

Type Aliases

MultipleRegistrationResponse 🔒

All responses voting key

SingleRegistrationResponse 🔒

Single registration response