Function cat_gateway::service::api::cardano::staked_ada_get::endpoint

pub(crate) async fn endpoint(
    state: &State,
    stake_address: StakeAddress,
    provided_network: Option<Network>,
    slot_num: Option<i64>
) -> UniResponse<Empty, Empty, Empty, OK<Json<StakeInfo>>, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, ApiValidationError, Empty, Empty, Empty, NotFound, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, ServerError, Empty, Empty, ServiceUnavailable, Empty, Empty, Empty, Empty, Empty, Empty, Empty>

GET /staked_ada