cat_gateway::service::api::cardano::staking

Module assets_get

Implementation of the GET ../assets endpoint

Structs

TxoAssetInfo 🔒

TXO asset information.

TxoAssetsState 🔒

TXO Assets state

TxoInfo 🔒

TXO information used when calculating a user’s stake info.

Enums

Responses 🔒

Endpoint responses.

Functions

build_full_stake_info_response 🔒

Building a full stake info response from the provided arguments.

build_stake_info 🔒

Builds an instance of StakeInfo based on the TXOs given.

calculate_assets_state 🔒

Calculate the assets state info for a given stake address.

endpoint 🔒

GET /staked_ada

get_stake_address_for_token 🔒

Fetches the stake address for a given Catalyst token.

get_stake_address_from_cat_id

Retrieves the CIP-19 stake address associated with a given Catalyst RBAC token.

get_txo 🔒

Returns a map of TXO infos for the given stake address.

get_txo_assets 🔒

Returns a map of txo asset infos for the given stake address.

update_spent 🔒

Checks if the given TXOs were spent and mark then as such. Separating base_txos and txos because we dont want to make an update inside the db for the base_txos data (it is covering the case when inside the persistent part we have a txo which is spent inside the volatile, so to not incorrectly mix up records from these two tables, inserting some rows from persistent to volatile section).

Type Aliases

AllResponses 🔒

All responses.

TxoAssetsMap 🔒

TXO Assets map type alias

TxoMap 🔒

TxoInfo map type alias