Function cat_gateway::service::api::cardano::sync_state_get::endpoint

pub(crate) async fn endpoint(
    state: &State,
    network: Option<Network>
) -> UniResponse<Empty, Empty, Empty, OK<Json<SyncState>>, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, ApiValidationError, Empty, Empty, Empty, NotFound, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, ServerError, Empty, Empty, ServiceUnavailable, Empty, Empty, Empty, Empty, Empty, Empty, Empty>

GET /sync_state