cat_gateway::service::api::cardano::types
Type Alias PublicVotingInfo

pub(crate) type PublicVotingInfo = VotingInfo;

Public voting key

Aliased Type

enum PublicVotingInfo {
    Direct(PubKey),
    Delegated(Vec<(PubKey, i64)>),
}

Variants

Direct(PubKey)

Direct voting

Voting power is based on the staked ada of the given key

Delegated(Vec<(PubKey, i64)>)

Delegated voting

Voting power is based on the staked ada of the delegated keys order of elements is important and must be preserved.