cat_gateway::service::api

Module documents

Signed Documents API endpoints

Modules

common 🔒

A module for placing common structs, functions, and variables across the document endpoint module not specified to a specific endpoint.

get_document 🔒

Implementation of the GET /document endpoint

post_document_index_query 🔒

Document Index Query

put_document 🔒

Implementation of the PUT /document endpoint

templates 🔒

Catalyst signed document templates.

Structs

DocumentApi 🔒

Cardano Follower API Endpoints