cat_gateway::service::api::documents

Module templates

Catalyst signed document templates.

Modules

data 🔒

Document templates hardcoded data.

Structs

SignedDocData 🔒

Signed document hardcoded data structure.

Constants

BRAND_ID 🔒

Catalyst brand ID.

BRAND_VERSION 🔒

Catalyst brand Version (same as ID).

CAMPAIGN_ID 🔒

Fund 14 Campaign ID.

CAMPAIGN_VERSION 🔒

Fund 14 Campaign Version (Same as ID).

Statics

TEMPLATES 🔒

A map of signed document templates to its ID.

Functions

build_signed_doc 🔒

Build a CatalystSignedDocument from the hardcoded SignedDocData, return a document with its id. This function should not fail, because template details are hardcoded.

get_doc_static_template 🔒

Get a static document template from ID and version.

load_doc_signing_key 🔒

Loads a Signing Key which is used to sign hardcoded Catalyst Signed Document objects