cat_gateway::service::api::documents::templates
Module data

Document templates list.

Constants

CATEGORY_DOC_TYPE 🔒

Category document type.

CATEGORY_TEMPLATES 🔒

List of category templates, 12 categories for Fund 14.

COMMENT_DOC_TYPE 🔒

Comment document type.

COMMENT_TEMPLATES 🔒

List of comment templates, 12 comments each of which is uniquely associated with one of the predefined categories.

PROPOSAL_DOC_TYPE 🔒

Proposal document type.

PROPOSAL_TEMPLATES 🔒

List of proposal templates, 12 proposals each of which is uniquely associated with one of the predefined categories.

Statics

TEMPLATE_DATA 🔒

List of raw templates, containing information to build signed document.