Function cat_gateway::service::api::test_endpoints::test_post::endpoint

pub(crate) async fn endpoint(
    id: i32,
    action: &String
) -> UniResponse<Empty, Empty, Empty, Empty, Empty, Empty, Empty, NoContent, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, ApiValidationError, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, ServerError, Empty, Empty, ServiceUnavailable, Empty, Empty, Empty, Empty, Empty, Empty, Empty>

POST /test/test

Liveness endpoint.

Kubernetes (and others) use this endpoint to determine if the service is able to keep running.

In this service, liveness is assumed unless there are multiple panics generated by an endpoint in a short window.

Responses

204 No Content - Service is OK and can keep running.

400 API Validation Error.

500 Server Error - If anything within this function fails unexpectedly. (Possible but unlikely)

503 Service Unavailable - Service is possibly not running reliably.