Function cat_gateway::service::api::utxo::staked_ada_get::endpoint

pub(crate) async fn endpoint(
    state: &State,
    stake_address: CardanoStakeAddress,
    provided_network: Option<Network>,
    date_time: Option<DateTime<Utc>>
) -> UniResponse<Empty, Empty, Empty, OK<Json<StakeInfo>>, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, ApiValidationError, Empty, Empty, Empty, NotFound, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, ServerError, Empty, Empty, ServiceUnavailable, Empty, Empty, Empty, Empty, Empty, Empty, Empty>

GET /utxo/staked_ada