Type Alias cat_gateway::service::api::v1::fragments_post::AllResponses

pub(crate) type AllResponses = UniResponse<Empty, Empty, Empty, OK<Json<FragmentsProcessingSummary>>, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, Empty, ServerError, Empty, Empty, ServiceUnavailable, Empty, Empty, Empty, Empty, Empty, Empty, Empty>;

All responses

Aliased Type

enum AllResponses {

    T100(Empty),
    T101(Empty),
    T102(Empty),
    T200(OK<Json<FragmentsProcessingSummary>>),
    T201(Empty),
    T202(Empty),
    T203(Empty),
    T204(Empty),
    T205(Empty),
    T206(Empty),
    T207(Empty),
    T208(Empty),
    T226(Empty),
    T300(Empty),
    T301(Empty),
    T302(Empty),
    T303(Empty),
    T304(Empty),
    T305(Empty),
    T307(Empty),
    T308(Empty),
    T400(Empty),
    T401(Empty),
    T402(Empty),
    T403(Empty),
    T404(Empty),
    T405(Empty),
    T406(Empty),
    T407(Empty),
    T408(Empty),
    T409(Empty),
    T410(Empty),
    T411(Empty),
    T412(Empty),
    T413(Empty),
    T414(Empty),
    T415(Empty),
    T416(Empty),
    T417(Empty),
    T418(Empty),
    T421(Empty),
    T422(Empty),
    T423(Empty),
    T424(Empty),
    T426(Empty),
    T428(Empty),
    T429(Empty),
    T431(Empty),
    T451(Empty),
    T500(ServerError),
    T501(Empty),
    T502(Empty),
    T503(ServiceUnavailable),
    T504(Empty),
    T505(Empty),
    T506(Empty),
    T507(Empty),
    T508(Empty),
    T510(Empty),
    T511(Empty),

}

Variants

T100(Empty)

T101(Empty)

T102(Empty)

T200(OK<Json<FragmentsProcessingSummary>>)

T201(Empty)

T202(Empty)

T203(Empty)

T204(Empty)

T205(Empty)

T206(Empty)

T207(Empty)

T208(Empty)

T226(Empty)

T300(Empty)

T301(Empty)

T302(Empty)

T303(Empty)

T304(Empty)

T305(Empty)

T307(Empty)

T308(Empty)

T400(Empty)

T401(Empty)

T402(Empty)

T403(Empty)

T404(Empty)

T405(Empty)

T406(Empty)

T407(Empty)

T408(Empty)

T409(Empty)

T410(Empty)

T411(Empty)

T412(Empty)

T413(Empty)

T414(Empty)

T415(Empty)

T416(Empty)

T417(Empty)

T418(Empty)

T421(Empty)

T422(Empty)

T423(Empty)

T424(Empty)

T426(Empty)

T428(Empty)

T429(Empty)

T431(Empty)

T451(Empty)

T500(ServerError)

T501(Empty)

T502(Empty)

T503(ServiceUnavailable)

T504(Empty)

T505(Empty)

T506(Empty)

T507(Empty)

T508(Empty)

T510(Empty)

T511(Empty)