cat_gateway::service::common::auth
Module api_key

API Key authorization scheme is used ONLY by internal endpoints.

Its purpose is to prevent their use externally, if they were accidentally exposed.

It is NOT to be used on any endpoint intended to be publicly facing.

Structs

InternalApiKeyAuthorization 🔒

ApiKey authorization for Internal Endpoints

Constants

API_KEY_HEADER 🔒

The header name that holds the API Key

Functions

api_checker 🔒

Check the provided API Key matches the API Key defined by for the deployment.

check_api_key 🔒

Check if the API Key is correctly set. Returns an error if it is not.