cat_gateway::service::common::auth::rbac

Module scheme

Catalyst RBAC Security Scheme definition Catalyst RBAC Security Scheme

Structs

AuthTokenAccessViolation

Token does not have required access rights

AuthTokenError

Error with the Authorization Token

CatalystRBACSecurityScheme

Catalyst RBAC Access Token

Constants

AUTHORIZATION_HEADER 🔒

The header name that holds the authorization RBAC token

MAX_TOKEN_AGE 🔒

Time in the past the Token can be valid for.

MAX_TOKEN_SKEW 🔒

Time in the future the Token can be valid for.

Statics

CACHE 🔒

Cached auth tokens

Functions

c509_key 🔒

Returns VerifyingKey from the given C509 certificate.

check_signature_algorithm 🔒

Checks that the signature algorithm is supported.

checker_api_catalyst_auth 🔒

When added to an endpoint, this hook is called per request to verify the bearer token is valid. The performed validation is described here.

indexed_registrations 🔒

Returns a list of all registrations for the given Catalyst ID from the database.

last_signing_key 🔒

Returns the last signing key from the registration chain.

registration 🔒

Returns a RBAC registration from the given block and slot.

registration_chain 🔒

Build a registration chain from the given indexed data.

service_unavailable 🔒

Returns a 503 error instance.

verifying_key 🔒

Creates VerifyingKey from the given extended public key.

x509_key 🔒

Returns VerifyingKey from the given X509 certificate.

Type Aliases

EncodedAuthToken

Auth token in the form of catv1..