Enum cat_gateway::service::common::objects::fragments_processing_summary::ReasonRejected

pub(crate) enum ReasonRejected {
    FragmentAlreadyInLog,
    FragmentInvalid,
    PreviousFragmentInvalid,
    PoolOverflow,
}

The reason for which a fragment was rejected.

Variants

FragmentAlreadyInLog

This fragment was already processed by the node.

FragmentInvalid

This fragment failed validation.

PreviousFragmentInvalid

One of the previous fragments was rejected and fail_fast is enabled.

PoolOverflow

One of the mempools rejected this fragment due to reaching capacity limit.