cat_gateway::service::common
Module types

Common types

These should be simple types, not objects. For example, types derived from strings or integers and vectors of simple types only.

Objects are objects, and not types.

Simple types can be enums, if the intended underlying type is simple, such as a string or integer.

Modules

cardano 🔒

Cardano Types

document 🔒

Signed Document Types

generic 🔒

Generic Types

headers 🔒

Header Types

payload 🔒

Payload Types

string_types 🔒

Simple string types.