cat_gateway::service::common::types::cardano
Module cip19_shelley_address

Cardano address types.

More information can be found in CIP-19

Structs

Cip19ShelleyAddress 🔒

Constants

BECH32 🔒

Bech32 Match Pattern

DECODED_STAKED_ADDR_LEN 🔒

Length of the decoded address.

DECODED_UNSTAKED_ADDR_LEN 🔒

Length of the decoded address.

DESCRIPTION 🔒

Description

ENCODED_STAKED_ADDR_LEN 🔒

Length of the encoded address (for type 0 - 3).

ENCODED_UNSTAKED_ADDR_LEN 🔒

Length of the encoded address (for type 6 - 7).

EXAMPLE 🔒

Example

MAX_LENGTH 🔒

Minimum length

MIN_LENGTH 🔒

Minimum length

PATTERN 🔒

Regex Pattern

PROD 🔒

Production Address Identifier

TEST 🔒

Test Address Identifier

TITLE 🔒

Title

Statics

EXTERNAL_DOCS 🔒

External document for Cardano addresses.

SCHEMA 🔒

Schema.

Functions

is_valid 🔒

Because ALL the constraints are defined above, we do not ever need to define them in the API. BUT we do need to make a validator. This helps enforce uniform validation.