cat_gateway::service::common::types::cardano

Module cip19_stake_address

Cardano address types.

More information can be found in CIP-19

Structs

Cip19StakeAddress 🔒

Constants

DECODED_ADDR_LEN 🔒

Length of the decoded address.

DESCRIPTION 🔒

Stake address description.

ENCODED_ADDR_LEN 🔒

Length of the encoded address.

EXAMPLE 🔒

Stake address example.

FORMAT 🔒

String Format

MAX_LENGTH 🔒

Minimum length

MIN_LENGTH 🔒

Minimum length

PATTERN 🔒

Regex Pattern

TITLE 🔒

Stake address title.

Statics

EXTERNAL_DOCS 🔒

External document for Cardano addresses.

STAKE_SCHEMA 🔒

Schema for StakeAddress.

Functions

is_valid 🔒

Because ALL the constraints are defined above, we do not ever need to define them in the API. BUT we do need to make a validator. This helps enforce uniform validation.