cat_gateway::service::common::types::cardano

Module query

These types are specifically and only used for Query Parameters

They exist due to limitations in the expressiveness of Query parameter constraints in OpenAPI

Modules

as_at 🔒

Query Parameter that can take either a Blockchain slot Number or Unix Epoch timestamp.

stake_or_voter 🔒

Query Parameter that can take a CIP-19 stake address, or a hex encoded vote public key.