cat_gateway::service::common::types::cardano::query

Module stake_or_voter

Query Parameter that can take a CIP-19 stake address, or a hex encoded vote public key.

Allows us to have one parameter that can represent two things, uniformly.

Structs

StakeOrVoter 🔒

Either a Stake Address or a ED25519 Public key.

Enums

StakeAddressOrPublicKey 🔒

A Query Parameter that can take a CIP-19 stake address, or a public key. Defining these are mutually exclusive, as a single parameter is required to be used.

Constants

DESCRIPTION 🔒

Description.

EXAMPLE 🔒

Example

FORMAT 🔒

Format

PATTERN 🔒

Validation Regex Pattern

STAKE_PATTERN 🔒

Stake Address Pattern

TITLE 🔒

Title.

VOTING_KEY_PATTERN 🔒

Voting Key Pattern

Statics

MAX_LENGTH 🔒

Maximum parameter length

MIN_LENGTH 🔒

Minimum parameter length

SCHEMA 🔒

Schema.