cat_gateway::service::common::types::document

Module id

Signed Document ID

UUIDv7 Encoded Document ID.

Structs

DocumentId 🔒

EqOrRangedIdDocumented 🔒

Document ID Selector

IdEq 🔒

A single Document IDs.

IdEqDocumented 🔒

ID Equals

IdRange 🔒

A range of Document IDs.

IdRangeDocumented 🔒

ID Range

Enums

EqOrRangedId 🔒

Either a Single Document ID, or a Range of Document IDs

Constants

DESCRIPTION 🔒

Description.

ENCODED_LENGTH 🔒

Length of the hex encoded string

EXAMPLE 🔒

Example.

EXAMPLE_MAX 🔒

Example maximum - Timestamp retained, random value set to all f

EXAMPLE_MIN 🔒

Example minimum - Timestamp retained, random value set to all 0

FORMAT 🔒

Format

PATTERN 🔒

Validation Regex Pattern

TITLE 🔒

Title.

URI 🔒

External Documentation URI

URI_DESCRIPTION 🔒

Description of the URI

Statics

SCHEMA 🔒

Schema

Functions

is_valid 🔒

Because ALL the constraints are defined above, we do not ever need to define them in the API. BUT we do need to make a validator. This helps enforce uniform validation.