cat_gateway::service::common::types::generic::query
Module pagination

Consistent Pagination Types

These types are paired and must be used together.

Page - The Page we wish to request, defaults to 0. Limit - The Limit we wish to request, defaults to 100.

Macros

limit_description 🔒

Description - must be suitable for both the Query and Response docs.

page_description 🔒

Description.

remaining_description 🔒

Description.

Structs

Limit 🔒

Limit of items to be returned in a page of data.

Page 🔒

Page to be returned in the response.

Remaining 🔒

Limit of items to be returned in a page of data.

Constants

LIMIT_DEFAULT 🔒

Default Limit (Should be used by paged responses to set the maximum size of the response).

LIMIT_DESCRIPTION 🔒

Description

LIMIT_EXAMPLE 🔒

Example.

LIMIT_MAXIMUM 🔒

Maximum.

LIMIT_MINIMUM 🔒

Minimum.

LIMIT_TITLE 🔒

Title.

PAGE_DEFAULT 🔒

Default

PAGE_DESCRIPTION 🔒

Description

PAGE_EXAMPLE 🔒

Example.

PAGE_MAXIMUM 🔒

Page Maximum.

PAGE_MINIMUM 🔒

Page Minimum.

PAGE_TITLE 🔒

Page Title.

REMAINING_DESCRIPTION 🔒

Description

REMAINING_EXAMPLE 🔒

Example.

REMAINING_MAXIMUM 🔒

Maximum.

REMAINING_MINIMUM 🔒

Minimum.

REMAINING_TITLE 🔒

Title.

Statics

LIMIT_SCHEMA 🔒

Schema.

PAGE_SCHEMA 🔒

Schema.

REMAINING_SCHEMA 🔒

Schema.

Functions

is_valid_limit 🔒

Is the Page valid?

is_valid_page 🔒

Is the Page valid?

is_valid_remaining 🔒

Is the Page valid?