cat_gateway::service::common::types::headers

Module retry_after

Retry After header type

This is an active header which expects to be provided in a response.

Enums

RetryAfterHeader 🔒

Parameter which describes the possible choices for a Retry-After header field.

RetryAfterOption 🔒

Parameter which lets us set the retry header, or use some default. Needed, because its valid to exclude the retry header specifically. This is also due to the way Poem handles optional headers.