cat_gateway::service::common::types::string_types
Macro impl_string_types

macro_rules! impl_string_types {
    ($(#[$docs:meta])* $ty:ident, $type_name:literal, $format:expr, $schema:expr ) => { ... };
    ($(#[$docs:meta])* $ty:ident, $type_name:literal, $format:expr, $schema:expr, $validator:expr) => { ... };
}

Macro to make creating validated and documented string types much easier.

Parameters

$ty - The Type name to create. Example MyNewType.

$type_name - The OpenAPI name for the type. Almost always going to be string.

$format - The OpenAPI format for the type. Where possible use a defined OpenAPI or JsonSchema format.

$schema - A Poem MetaSchema which defines all the schema parameters for the type.

$validation - OPTIONAL Validation function to apply to the string value.

Example

impl_string_types!(MyNewType, "string", "date", MyNewTypeSchema, SomeValidationFunction);

Is the equivalent of:

#[derive(Debug, Clone, Eq, PartialEq, Hash)]
pub(crate) struct MyNewType(pub String);

impl <stuff> for MyNewType { ... }