Module cat_gateway::service::utilities

API Utility operations

Modules

catch_panic 🔒

Handle catching panics created by endpoints, logging them and properly responding.

middleware 🔒

Custom POEM Middleware for this service.

Functions

check_network 🔒

Check the provided network type with the encoded inside the stake address

to_hex_with_prefix 🔒

Convert bytes to hex string with the 0x prefix