cat_gateway::service::utilities

Module convert

Simple general purpose utility functions.

Functions

from_saturating 🔒

Convert an <T> to <R>. (saturate if out of range.) Note can convert any int to float, or f32 to f64 as well. can not convert from float to int, or f64 to f32.