cat_gateway::service::utilities

Module health

Utilities used for health functionality.

Statics

IS_LIVE 🔒

Flag to determine if the service is live.

LIVE_COUNTER 🔒

Timestamp in seconds used to determine if the service is live.

Functions

get_current_timestamp 🔒

Returns the timestamp of current system time, in seconds.

get_live_counter 🔒

Get the LIVE_COUNTER timestamp value, in seconds.

is_live 🔒

Get the IS_LIVE flag.

set_live_counter 🔒

Set the LIVE_COUNTER to current timestamp, in seconds.

set_not_live 🔒

Set the IS_LIVE flag to false.