cat_gateway::service::utilities::health

Module start

Utilities used for system Started functionality.

Statics

INITIAL_FOLLOWER_TIP_REACHED 🔒

Flag to determine if the chain follower has synchronized with the Tip for the first time.

LIVE_EVENT_DB 🔒

Flag to determine if the Event DB has started.

LIVE_INDEX_DB 🔒

Flag to determine if the Indexing DB has started.

STARTED 🔒

Flag to determine if the service has started

Functions

condition_for_started 🔒

Returns whether the service has started or not.

event_db_is_live 🔒

Returns whether the Event DB is live or not.

follower_has_first_reached_tip 🔒

Returns whether the Chain Follower has reached Tip or not.

index_db_is_live 🔒

Returns whether the Index DB is live or not.

service_has_started 🔒

Returns whether the service has started or not.

set_event_db_liveness 🔒

Set the LIVE_EVENT_DB flag.

set_follower_first_reached_tip 🔒

Set the INITIAL_FOLLOWER_TIP_REACHED as true.

set_index_db_liveness 🔒

Set the LIVE_INDEX_DB flag.

set_to_started 🔒

Set the STARTED flag to true