cat_gateway::service::utilities

Module middleware

Custom POEM Middleware for this service.

Modules

db_check 🔒

Middleware to verify that there is connection the databases.

schema_validation 🔒

Middleware to verify the status of the last DB schema version validation.

tracing_mw 🔒

Full Tracing and metrics middleware.