Module cat_gateway::service::utilities::middleware::tracing_mw

Full Tracing and metrics middleware.

Structs

CLIENT_REQUEST_COUNT 🔒

HTTP_REQUEST_COUNT 🔒

HTTP_REQ_CPU_TIME_MS 🔒

HTTP_REQ_DURATION_MS 🔒

PANIC_REQUEST_COUNT 🔒

ResponseData 🔒

Data we collected about the response

Tracing 🔒

Middleware for tracing.

TracingEndpoint 🔒

Endpoint for Tracing middleware.

Constants

CLIENT_METRIC_LABELS 🔒

Labels for the client metrics

METRIC_LABELS 🔒

Labels for the metrics

Functions

add_interesting_headers_to_span 🔒

Add all interesting headers to the correct fields in a span. This logic is the same for both requests and responses.

anonymous_client_id 🔒

Get an anonymized client ID from the request.

init_prometheus 🔒

Initialize Prometheus metrics.

mk_request_span 🔒

Make a span from the request