Module cat_gateway::settings

Command line and environment variable settings for the service

Structs

CassandraEnvVars 🔒

Configuration for an individual cassandra cluster.

DocsSettings 🔒

Settings specifies OpenAPI docs generation.

EnvVars 🔒

All the EnvVars used by the service.

FollowerSettings 🔒

Settings for follower mechanics.

ServiceSettings 🔒

Settings for the application.

Settings 🔒

Our Global Settings for this running service.

StringEnvVar 🔒

An environment variable read as a string.

Constants

ADDRESS_DEFAULT 🔒

Default address to start service on.

API_HOST_NAMES_DEFAULT 🔒

Default API_HOST_NAME/S used in production. This can be a single hostname, or a list of them.

API_URL_PREFIX_DEFAULT 🔒

Default API_URL_PREFIX used in development.

CASSANDRA_PERSISTENT_DB_NAMESPACE_DEFAULT 🔒

Default Cassandra DB URL for the Persistent DB.

CASSANDRA_PERSISTENT_DB_URL_DEFAULT 🔒

Default Cassandra DB URL for the Persistent DB.

CASSANDRA_VOLATILE_DB_NAMESPACE_DEFAULT 🔒

Default Cassandra DB URL for the Persistent DB.

CASSANDRA_VOLATILE_DB_URL_DEFAULT 🔒

Default Cassandra DB URL for the Persistent DB.

CHECK_CONFIG_TICK_DEFAULT 🔒

Default CHECK_CONFIG_TICK used in development.

CLIENT_ID_KEY_DEFAULT 🔒

Default CLIENT_ID_KEY used in development.

EVENT_DB_URL_DEFAULT 🔒

Default Event DB URL.

GITHUB_ISSUE_TEMPLATE_DEFAULT 🔒

Default Github issue template to use

GITHUB_REPO_NAME_DEFAULT 🔒

Default Github repo name

GITHUB_REPO_OWNER_DEFAULT 🔒

Default Github repo owner

MACHINE_UID_DEFAULT 🔒

Default MACHINE_UID used in development

Statics

ENV_VARS 🔒

Handle to the mithril sync thread. One for each Network ONLY.

SERVICE_SETTINGS 🔒

All Settings/Options for the Service.

Functions

calculate_service_uuid 🔒

Hash the Public IPv4 and IPv6 address of the machine, and convert to a 128 bit V4 UUID.

cassandra_cfg 🔒

Create a config for a cassandra cluster, identified by a default namespace.

string_to_api_host_names 🔒

Transform a string list of host names into a vec of host names. Default to the service address if none specified.