cat_gateway::settings
Module cassandra_db

Command line and environment variable settings for the service

Structs

EnvVars 🔒

Configuration for an individual cassandra cluster.

Constants

MAX_BATCH_SIZE 🔒

Maximum possible batch size.

MAX_BATCH_SIZE_DEFAULT 🔒

Default maximum batch size. This comes from: https://docs.aws.amazon.com/keyspaces/latest/devguide/functional-differences.html#functional-differences.batch Scylla may support larger batches for better performance. Larger batches will incur more memory overhead to store the prepared batches.

MIN_BATCH_SIZE 🔒

Minimum possible batch size.

PERSISTENT_NAMESPACE_DEFAULT 🔒

Default Cassandra DB URL for the Persistent DB.

PERSISTENT_URL_DEFAULT 🔒

Default Cassandra DB URL for the Persistent DB.

VOLATILE_NAMESPACE_DEFAULT 🔒

Default Cassandra DB URL for the Persistent DB.

VOLATILE_URL_DEFAULT 🔒

Default Cassandra DB URL for the Persistent DB.