cat_gateway::settings

Struct Settings

pub(crate) struct Settings();

Our Global Settings for this running service.

Implementations

impl Settings

pub(crate) fn init(settings: ServiceSettings) -> Result<()>

Initialize the settings data.

pub(crate) fn event_db_settings() -> (&'static str, Option<&'static str>, Option<&'static str>)

Get the current Event DB settings for this service.

pub(crate) fn cassandra_db_cfg() -> (EnvVars, EnvVars)

Get the Persistent & Volatile Cassandra DB config for this service.

pub(crate) fn follower_cfg() -> EnvVars

Get the configuration of the chain follower.

pub(crate) fn cardano_network() -> Network

Chain Follower network (The Blockchain network we are configured to use). Note: Catalyst Gateway can ONLY follow one network at a time.

pub(crate) fn api_url_prefix() -> &'static str

The API Url prefix

pub(crate) fn client_id_key() -> &'static str

The Key used to anonymize client connections in the logs.

pub(crate) fn service_id() -> &'static str

The Service UUID

pub(crate) fn metrics_memory_interval() -> Duration

The memory metrics interval

pub(crate) fn metrics_follower_interval() -> Duration

The Chain Follower metrics interval

pub(crate) fn api_host_names() -> Vec<String>

Get a list of all host names to serve the API on.

Used by the OpenAPI Documentation to point to the correct backend. Take a list of [scheme://] + host names from the env var and turns it into a lits of strings.

Host names are taken from the API_HOST_NAMES environment variable. If that is not set, returns an empty list.

pub(crate) fn bound_address() -> SocketAddr

The socket address we are bound to.

pub(crate) fn server_name() -> Option<&'static str>

Get the server name to be used in the Server object of the OpenAPI Document.

pub(crate) fn generate_github_issue_url(title: &str) -> Option<Url>

Generate a github issue url with a given title

Arguments

title: &str - the title to give the issue

Returns

String - the url

Example

assert_eq!(
    generate_github_issue_url("Hello, World! How are you?"),
    "https://github.com/input-output-hk/catalyst-voices/issues/new?template=bug_report.yml&title=Hello%2C%20World%21%20How%20are%20you%3F"
);

pub(crate) fn check_internal_api_key(value: &str) -> bool

Check a given key matches the internal API Key

pub(crate) fn purge_slot_buffer() -> Slot

Slot buffer used as overlap when purging Live Index data.

pub(crate) fn service_live_timeout_interval() -> Duration

Duration in seconds used to determine if the system is live during checks.

pub(crate) fn service_live_counter_threshold() -> u64

Value after which the service is considered NOT live.