cat_gateway::utils::blake2b_hash

Function generate_uuid_string_from_data

pub(crate) fn generate_uuid_string_from_data(
    key: &str,
    data: &[String],
) -> String

Generates a UUID string from the provided key and data using BLAKE2b hashing.

Arguments

key: A string slice that is used as part of the hash function input.

data: A vector of strings which will be included in the BLAKE2b hash computation.

Returns

A UUID string generated from the BLAKE2b hash of the concatenated data with the key.