Module cat_gateway::cardano

Logic for orchestrating followers

Modules

cip36_registration 🔒

Verify registration TXs

util 🔒

Block stream parsing and filtering utils

Constants

MAX_BLOCKS_BATCH_LEN 🔒

Blocks batch length that will trigger the blocks buffer to be written to the database.

Functions

get_follower_config 🔒

Returns a follower configs, waits until they present inside the db

index_block_buffer 🔒

Consumes a block buffer and indexes its data.

index_blocks 🔒

Index the data from the given blocks.

index_many_blocks 🔒

Index a slice of blocks.

index_transaction_inputs_data 🔒

Index transaction inputs data.

index_transaction_outputs_data 🔒

Index transaction outputs data.

index_transactions 🔒

Index transactions (and its inputs and outputs) from a slice of blocks.

index_transactions_data 🔒

Index transactions data.

index_voter_registrations 🔒

Index voter registrations from a slice of blocks.

instantiate_follower 🔒

Instantiate the follower. If there is metadata present which allows us to bootstrap from a point in time We start from there, if not; we start from genesis.

process_blocks 🔒

Process next block from the follower

spawn_follower 🔒

Initiate single follower and returns associated task handler which facilitates future control over spawned threads.

spawn_followers 🔒

Spawn follower threads and return associated handlers

start_followers 🔒

Start followers as per defined in the config

Type Aliases

ManageTasks

Handler for follower tasks, allows for control over spawned follower threads