cat_gateway::db

Module event

Catalyst Election Database crate

Modules

common 🔒

Reusable common database objects

config 🔒

Configuration query

error 🔒

Database Errors

schema_check 🔒

Check if the schema is up-to-date.

signed_docs 🔒

Signed docs queries

Structs

EventDB 🔒

The Catalyst Event SQL Database

Enums

EventDBConnectionError 🔒

EventDB Errors

Constants

DATABASE_SCHEMA_VERSION 🔒

Database version this crate matches. Must equal the last Migrations Version Number from event-db/migrations.

Statics

DEEP_QUERY_INSPECT 🔒

Is Deep Query Analysis enabled or not?

EVENT_DB_POOL 🔒

Postgres Connection Manager DB Pool Instance

Functions

establish_connection_pool

Establish a connection to the database, and check the schema is up-to-date.

Type Aliases

SqlDbPool 🔒

Postgres Connection Manager DB Pool