Module cat_gateway::db::index::schema

Index Schema

Constants

CREATE_NAMESPACE_CQL 🔒

Keyspace Create (Templated)

SCHEMAS 🔒

All Schema Creation Statements

SCHEMA_VERSION 🔒

The version of the Schema we are using. Must be incremented if there is a breaking change in any schema tables below.

Functions

create_namespace 🔒

Create the namespace we will use for this session Ok to run this if the namespace already exists.

create_schema 🔒

Create the Schema on the connected Cassandra DB

namespace 🔒

Get the namespace for a particular db configuration