cat_gateway::db::index
Module schema

Index Schema

Constants

CREATE_NAMESPACE_CQL 🔒

Keyspace Create (Templated)

SCHEMAS 🔒

All Schema Creation Statements

Functions

create_namespace 🔒

Create the namespace we will use for this session Ok to run this if the namespace already exists.

create_schema 🔒

Create the Schema on the connected Cassandra DB

generate_cql_schema_version 🔒

Generates a unique schema version identifier based on the content of all CQL schemas.

namespace 🔒

Get the namespace for a particular db configuration

remove_comments_and_join_query_lines 🔒

Removes all comments from each line in the input query text and joins the remaining lines into a single string, reducing consecutive whitespace characters to a single space. Comments are defined as any text following -- on a line.