cat_gateway::metrics::chain_indexer

Module reporter

All the related Chain Indexer reporting metrics to the Prometheus service are inside this module.

Constants

METRIC_LABELS 🔒

Labels for the metrics.

Statics

CURRENT_IMMUTABLE_TIP_SLOT 🔒

Chain Indexer current immutable tip slot.

CURRENT_LIVE_TIP_SLOT 🔒

Chain Indexer current live tip slot.

HIGHEST_COMPLETE_INDEXED_SLOT 🔒

Chain Indexer highest complete indexed slot.

IMMUTABLE_REACHED_TIP 🔒

Chain Indexer indicator whether immutable tip is reached or in progress.

LIVE_REACHED_TIP 🔒

Chain Indexer indicator whether live tip is reached or in progress.

PURGED_SLOTS 🔒

Chain Indexer number of purged slots.

RUNNING_INDEXER_TASKS_COUNT 🔒

Chain Indexer number of current running indexer tasks.

SLOT_TIP_DIFF 🔒

The slot difference between immutable tip slot and the volatile tip slot number.

TRIGGERED_BACKWARD_PURGES_COUNT 🔒

Chain Indexer number of times triggering backward data purge.

TRIGGERED_FORWARD_PURGES_COUNT 🔒

Chain Indexer number of times triggering forward data purge.