cat_gateway::service

Module api

Catalyst Gateway API Definition

This defines all endpoints for the Catalyst Gateway API. It however does NOT contain any processing for them, that is defined elsewhere.

Modules

cardano 🔒

Cardano API endpoints

config 🔒

Configuration Endpoints

documents 🔒

Signed Documents API endpoints

health 🔒

Health Endpoints

Constants

API_DESCRIPTION 🔒

A long description of the API. Markdown is supported

API_TITLE 🔒

The name of the API

API_VERSION 🔒

The version of the API

TERMS_OF_SERVICE 🔒

Get the terms of service for the API

Functions

get_api_contact 🔒

Get the contact details for inquiring about the API

get_api_license 🔒

Get the license details for the API

mk_api 🔒

Create the OpenAPI definition

set_localhost_addresses 🔒

Set the localhost addresses descriptions.