cat_gateway::service::api::cardano::cip36

Module filter

Implementation of the GET /cardano/cip36 endpoint

Functions

build_response 🔒

Build a final response which will contain all found registrations, sort them and apply pagination.

get_all_invalid_registrations 🔒

Get all invalid registrations

get_all_valid_registrations 🔒

Get all valid cip36 registrations.

get_invalid_registrations 🔒

Get invalid registrations for stake public key

get_registrations_given_stake_addr 🔒

Get registrations given a stake address, it can be time specific based on asat param, or the all registrations returned if no asat given.

get_registrations_given_vote_key 🔒

Get registrations given a vote key, time specific based on asat param, or latest registration returned if no asat given.

get_valid_registrations 🔒

Get valid cip36 registrations for a given stake public key.

paginate_registrations 🔒

Paginate the registrations.

snapshot 🔒

Get all registrations or constrain if slot# given.

sort_registrations 🔒

Sort registrations by slot number, nonce, and transaction offset. If slot_no is the same, the registration with the highest nonce wins. If nonce is the same, the registration with the highest txn_offset wins.