cat_gateway::service::api::documents::templates

Module data

Document templates hardcoded data.

Structs

CategoryDocData 🔒

A definition of Category document hardcoded data

CommentTemplateDocData 🔒

A definition of Comment Template document hardcoded data

ProposalTemplateDocData 🔒

A definition of Proposal Template document hardcoded data

Constants

CATEGORY_DOCUMENTS 🔒

List of category documents, 12 categories for Fund 14.

COMMENT_TEMPLATES 🔒

List of comment templates, 12 comments each of which is uniquely associated with one of the predefined categories.

EMPTY_JSON_OBJECT_BYTES 🔒

An empty JSON object bytes slice

PROPOSAL_TEMPLATES 🔒

List of proposal templates, 12 proposals each of which is uniquely associated with one of the predefined categories.