cat_gateway::service::common

Module responses

Generic Responses are all contained in their own modules, grouped by response codes.

Modules

code_401_unauthorized 🔒

Define Unauthorized response type.

code_403_forbidden 🔒

Define Forbidden response type.

code_412_precondition_failed 🔒

Define Precondition Failed response type.

code_429_too_many_requests 🔒

Define TooManyRequests response type.

code_500_internal_server_error 🔒

Define ServerError type.

code_503_service_unavailable 🔒

Define Service Unavailable Response Body.

Structs

FilteredByStatusCodeResponse 🔒

FilteredByStatusCodeResponse is used to filter out duplicate responses by status code.

Enums

ErrorResponses 🔒

Default error responses

WithErrorResponses 🔒

Combine provided responses type with the default responses under one type.