cat_gateway::service::utilities::middleware

Module tracing_mw

Full Tracing and metrics middleware.

Structs

ResponseData 🔒

Data we collected about the response

Tracing 🔒

Middleware for tracing.

TracingEndpoint 🔒

Endpoint for Tracing middleware.

Functions

add_interesting_headers_to_span 🔒

Add all interesting headers to the correct fields in a span. This logic is the same for both requests and responses.

anonymize_ip_address 🔒

Given a Clients IP Address, return the anonymized version of it.

anonymous_client_id 🔒

Get an anonymized client ID from the request.

mk_request_span 🔒

Make a span from the request