Module cat_gateway::settings

Command line and environment variable settings for the service

Structs

CassandraEnvVars 🔒

Configuration for an individual cassandra cluster.

ChainFollowerEnvVars 🔒

Configuration for the chain follower.

DocsSettings 🔒

Settings specifies OpenAPI docs generation.

EnvVars 🔒

All the EnvVars used by the service.

ServiceSettings 🔒

Settings for the application.

Settings 🔒

Our Global Settings for this running service.

StringEnvVar 🔒

An environment variable read as a string.

Enums

StringEnvVarParams 🔒

Ergonomic way of specifying if a env var needs to be redacted or not.

Constants

ADDRESS_DEFAULT 🔒

Default address to start service on.

API_HOST_NAMES_DEFAULT 🔒

Default API_HOST_NAME/S used in production. This can be a single hostname, or a list of them.

API_URL_PREFIX_DEFAULT 🔒

Default API_URL_PREFIX used in development.

CASSANDRA_MAX_BATCH_SIZE 🔒

Maximum possible batch size.

CASSANDRA_MAX_BATCH_SIZE_DEFAULT 🔒

Default maximum batch size. This comes from: https://docs.aws.amazon.com/keyspaces/latest/devguide/functional-differences.html#functional-differences.batch Scylla may support larger batches for better performance. Larger batches will incur more memory overhead to store the prepared batches.

CASSANDRA_MIN_BATCH_SIZE 🔒

Minimum possible batch size.

CASSANDRA_PERSISTENT_DB_NAMESPACE_DEFAULT 🔒

Default Cassandra DB URL for the Persistent DB.

CASSANDRA_PERSISTENT_DB_URL_DEFAULT 🔒

Default Cassandra DB URL for the Persistent DB.

CASSANDRA_VOLATILE_DB_NAMESPACE_DEFAULT 🔒

Default Cassandra DB URL for the Persistent DB.

CASSANDRA_VOLATILE_DB_URL_DEFAULT 🔒

Default Cassandra DB URL for the Persistent DB.

CHAIN_FOLLOWER_DEFAULT 🔒

Default chain to follow.

CHAIN_FOLLOWER_SYNC_TASKS_DEFAULT 🔒

Default number of sync tasks (must be in the range 1 to 255 inclusive.)

CHECK_CONFIG_TICK_DEFAULT 🔒

Default CHECK_CONFIG_TICK used in development.

CLIENT_ID_KEY_DEFAULT 🔒

Default CLIENT_ID_KEY used in development.

EVENT_DB_URL_DEFAULT 🔒

Default Event DB URL.

GITHUB_ISSUE_TEMPLATE_DEFAULT 🔒

Default Github issue template to use

GITHUB_REPO_NAME_DEFAULT 🔒

Default Github repo name

GITHUB_REPO_OWNER_DEFAULT 🔒

Default Github repo owner

Statics

ENV_VARS 🔒

Handle to the mithril sync thread. One for each Network ONLY.

SERVICE_SETTINGS 🔒

All Settings/Options for the Service.

Functions

calculate_service_uuid 🔒

Hash the Public IPv4 and IPv6 address of the machine, and convert to a 128 bit V4 UUID.

string_to_api_host_names 🔒

Transform a string list of host names into a vec of host names. Default to the service address if none specified.