cat_gateway
Module settings

Command line and environment variable settings for the service

Modules

cassandra_db 🔒

Command line and environment variable settings for the service

chain_follower 🔒

Command line and environment variable settings for the service

str_env_var 🔒

Processing for String Environment Variables

Structs

DocsSettings 🔒

Settings specifies OpenAPI docs generation.

EnvVars 🔒

All the EnvVars used by the service.

ServiceSettings 🔒

Settings for the application.

Settings 🔒

Our Global Settings for this running service.

Constants

ADDRESS_DEFAULT 🔒

Default address to start service on.

API_HOST_NAMES_DEFAULT 🔒

Default API_HOST_NAME/S used in production. This can be a single hostname, or a list of them.

API_URL_PREFIX_DEFAULT 🔒

Default API_URL_PREFIX used in development.

CHECK_CONFIG_TICK_DEFAULT 🔒

Default CHECK_CONFIG_TICK used in development.

CLIENT_ID_KEY_DEFAULT 🔒

Default CLIENT_ID_KEY used in development.

EVENT_DB_URL_DEFAULT 🔒

Default Event DB URL.

GITHUB_ISSUE_TEMPLATE_DEFAULT 🔒

Default Github issue template to use

GITHUB_REPO_NAME_DEFAULT 🔒

Default Github repo name

GITHUB_REPO_OWNER_DEFAULT 🔒

Default Github repo owner

Statics

ENV_VARS 🔒

Handle to the mithril sync thread. One for each Network ONLY.

SERVICE_SETTINGS 🔒

All Settings/Options for the Service.

Functions

calculate_service_uuid 🔒

Hash the Public IPv4 and IPv6 address of the machine, and convert to a 128 bit V4 UUID.

string_to_api_host_names 🔒

Transform a string list of host names into a vec of host names. Default to the service address if none specified.