Module cat_gateway::settings

Command line and environment variable settings for the service

Structs

API_HOST_NAMES 🔒

A List of servers to provideThe client id key used to anonymize client connections.

API_URL_PREFIX 🔒

The base path the API is served at.

CHECK_CONFIG_TICK 🔒

Tick every N seconds until config exists in db

CLIENT_ID_KEY 🔒

The client id key used to anonymize client connections.

DocsSettings 🔒

Settings specifies OpenAPI docs generation.

FollowerSettings 🔒

Settings for follower mechanics.

GITHUB_ISSUE_TEMPLATE 🔒

The github issue template to use

GITHUB_REPO_NAME 🔒

The github repo name

GITHUB_REPO_OWNER 🔒

The github repo owner

ServiceSettings 🔒

Settings for the application.

StringEnvVar 🔒

An environment variable read as a string.

Constants

ADDRESS_DEFAULT 🔒

Default address to start service on.

API_HOST_NAMES_DEFAULT 🔒

Default API_HOST_NAME/S used in production. This can be a single hostname, or a list of them.

API_URL_PREFIX_DEFAULT 🔒

Default API_URL_PREFIX used in development.

CHECK_CONFIG_TICK_DEFAULT 🔒

Default CHECK_CONFIG_TICK used in development.

CLIENT_ID_KEY_DEFAULT 🔒

Default CLIENT_ID_KEY used in development.

DATA_REFRESH_TICK_DEFAULT 🔒

Default DATA_REFRESH_TICK used in development

GITHUB_ISSUE_TEMPLATE_DEFAULT 🔒

Default Github issue template to use

GITHUB_REPO_NAME_DEFAULT 🔒

Default Github repo name

GITHUB_REPO_OWNER_DEFAULT 🔒

Default Github repo owner

MACHINE_UID_DEFAULT 🔒

Default MACHINE_UID used in development

Functions

generate_github_issue_url 🔒

Generate a github issue url with a given title

get_api_host_names 🔒

Get a list of all host names to serve the API on.

string_to_api_host_names 🔒

Transform a string list of host names into a vec of host names. Default to the service address if none specified.