cat_gateway::settings

Module chain_follower

Command line and environment variable settings for the service

Structs

EnvVars 🔒

Configuration for the chain follower.

Constants

DEFAULT_NETWORK 🔒

Default chain to follow.

DEFAULT_SYNC_MAX_SLOTS 🔒

Default number of slots each sync task will process at one time. This default is just over one week worth of data where 1 slot == 1 second.

DEFAULT_SYNC_TASKS 🔒

Default number of sync tasks (must be in the range 1 to 256 inclusive.)

MAX_DL_CHUNK_QUEUE_AHEAD 🔒

Maximum DL Chunk Queue Ahead (must be in the range 1 to 256 inclusive.)

MAX_DL_CHUNK_SIZE 🔒

Maximum DL Chunk Size in MB (must be in the range 1 to 256 inclusive.)

MAX_DL_CONNECTIONS 🔒

Maximum number of DL Connections (must be in the range 1 to 256 inclusive.)

MAX_DL_TIMEOUT 🔒

Maximum DL Chunk Connect/Data Timeout in seconds (0 = Disabled).

MAX_SYNC_MAX_SLOTS 🔒

Maximum the number of slots each sync task will process at one time can be set to.

MAX_SYNC_TASKS 🔒

Maximum number of sync tasks (must be in the range 1 to 256 inclusive.)

MIN_SYNC_MAX_SLOTS 🔒

Minimum the number of slots each sync task will process at one time can be set to. Note: This is just the setting minimum, a sync task may sync as few as a 1 slot.

ONE_MEGABYTE 🔒

Number of bytes in a Megabyte